log: prepare trace tests for better testing
Release justification: testing improvement
Release note: None
