+++ pkg/util/log/BUILD.bazel
@@ -132,9 +132,11 @@ go_test(
         "fluent_client_test.go",
         "format_crdb_v2_test.go",
         "format_json_test.go",
+        "helpers_test.go",
         "main_test.go",
         "redact_test.go",
         "secondary_log_test.go",
+        "trace_client_test.go",
         "trace_test.go",
     ],
     data = glob(["testdata/**"]),
