+++ pkg/util/log/helpers_test.go
@@ -1,0 +1,17 @@
+// Copyright 2021 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package log
+
+// NoLogV returns a verbosity level that will not result in VEvents and
+// VErrEvents being logged.
+func NoLogV() Level {
+	return logging.vmoduleConfig.verbosity.get() + 1
+}
